rename consts.rs -> limits.rs
