+++ rust/protocol/src/group_cipher.rs
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 //
 
-use crate::consts::limits::MAX_FORWARD_JUMPS;
+use crate::limits::MAX_FORWARD_JUMPS;
 use crate::protocol::SENDERKEY_MESSAGE_CURRENT_VERSION;
 use crate::sender_keys::{SenderKeyState, SenderMessageKey};
 use crate::{
